requirements.txt
python-dotenv
openai
pyupbit
google-genai
ta